Jenkins: Add multiple unfinished jobs feature
I found it quite easy for mergeatron to get stuck in a state where the
first "unfinished" job in the DB had been removed from (or never got
added to) jenkins.  This meant that mergeatron was completely stuck as
it was always only checking for the first unfinished job and never got
past it.

This simply loops over the unfinished jobs performing the same actions
as before.

After adding this, mergeatron processed the entire backlog of 10-20 jobs
very quickly and kept on moving.